backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
# clerk-sdk-python
clerk-backend-api
langchain
langchain-ai21
python-dotenv
langchain-community
langchain_ollama